+++ poc/ConsoleHost/Program.cs
@@ -44,29 +44,4 @@ namespace Wcf.Extensions.OpenIdConnect.Poc.ConsoleHost
             return binding;
         }
     }
-
-    [ServiceContract]
-    public interface IService
-    {
-        [OperationContract]
-        string Ping();
-    }
-
-    // Init option 3: Add behavior by service attribute
-    [WrappedJwtAuthorizationServiceBehavior(requiredScopes: "write")]
-    internal class Service : IService
-    {
-        public string Ping()
-        {
-            var sb = new StringBuilder();
-            sb.AppendLine("Pong");
-
-            foreach (var claim in ClaimsPrincipal.Current.Claims)
-            {
-                sb.AppendFormat("{0} :: {1}\n", claim.Type, claim.Value);
-            }
-
-            return sb.ToString().Trim();
-        }
-    }
 }
