+++ poc/WebHost/Web.config
@@ -43,7 +43,8 @@
     </protocolMapping>
     <serviceHostingEnvironment>
       <serviceActivations>
-        <add service="Wcf.Extensions.OpenIdConnect.Poc.WebHost.Service" relativeAddress="./do-stuff.svc"
+        <add relativeAddress="do-stuff.svc"
+             service="Wcf.Extensions.OpenIdConnect.Poc.WebHost.Service"
              factory="System.ServiceModel.Activation.ServiceHostFactory" />
       </serviceActivations>
     </serviceHostingEnvironment>
